refactor(deps): move PluralRules abstract operations to ecma402-abstract
Move 4 ECMA-402 Section 16 abstract operations from
intl-pluralrules/abstract/ to ecma402-abstract/PluralRules/:
- GetOperands, InitializePluralRules, ResolvePlural, ResolvePluralRange

Fix type errors in InitializePluralRules for notation/compactDisplay
extension options (formatjs-specific, not in PluralRulesOptions type).
